requirements.txt
cantools
tensorflow
pandas
scikit-learn
matplotlib
scipy
pandas